fiware-datamodel-harvester/requirements.txt
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
genson==1.2.2
gitdb==4.0.9
GitPython==3.1.26
idna==3.3
importlib-resources==5.4.0
json-ref-dict==0.7.1
jsonpointer==2.2
jsonschema==4.4.0
mysql-connector-python==8.0.28
protobuf==3.19.4
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
rfc3987==1.3.8
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
statham-schema==0.13.5
typing-extensions==3.10.0.2
urllib3==1.26.8
zipp==3.7.0
flask==2.1.3
